+++ packages/ui/CHANGELOG.md
@@ -3,6 +3,15 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.86.1"></a>
+## [0.86.1](https://github.com/ionic-team/stencil-component-starter/compare/v0.86.0...v0.86.1) (2019-01-03)
+
+**Note:** Version bump only for package @bearer/ui
+
+
+
+
+
 <a name="0.86.0"></a>
 # [0.86.0](https://github.com/ionic-team/stencil-component-starter/compare/v0.85.4...v0.86.0) (2019-01-03)